feat(db): MetricRepository trait + PgMetricRepository
Adds the persistence layer for metric_definitions:

- `MetricRepository` trait in stitchd-db::repository (mod.rs)
  - find_by_id / find_by_key / find_batch_by_ids
  - list_by_environment / list_by_environment_paginated (COUNT(*) OVER())
  - create / update (optimistic-locking via version) / soft_delete
- `PgMetricRepository` in repository/pg/metric.rs
  - Raw `sqlx::query` strings (no `.sqlx` cache impact for parallel waves)
  - Audit-log integration via AuditLogger
  - `(kind TEXT, config JSONB)` column split is hidden behind row_to_metric
    helper — domain MetricKind is reconstructed from the split-shape by
    re-tagging the JSONB with the `kind` discriminator

Also fixes `MetricDefinition.version` from i32 → i64 to match the
`BIGINT` migration column and the existing pattern used by
`EventDefinition`, `Segment`, etc.

17 integration tests via `#[sqlx::test(migrations = "./migrations")]`:
- CRUD round-trip for each MetricKind variant (aggregation/ratio/funnel)
- find_by_id/find_by_key NotFound paths
- Unique violation on duplicate (env_id, key)
- Soft-delete enables recreate-with-same-key (partial UNIQUE index)
- list_by_environment filters soft-deleted; paginated returns total
- update_stale_version returns VersionConflict
- find_batch_by_ids filters soft-deleted + empty input

Wired:
- `pub mod metric` + `pub use metric::PgMetricRepository` in
  repository/pg/mod.rs
- `MetricRepository` + `PgMetricRepository` re-exported from crate root

phase1_task3 of events_metrics_20260519 (beads feature-flag-7an.1.3).

+++ crates/stitchd-core/src/metric/mod.rs
@@ -75,7 +75,9 @@ pub struct MetricDefinition {
     pub goal_direction: GoalDirection,
     /// Optimistic-locking version — incremented by the repo on every
     /// successful update; updates with stale `version` are rejected.
-    pub version: i32,
+    /// Stored as `BIGINT` in PG; matches the convention used by
+    /// `EventDefinition`, `Segment`, and other versioned entities.
+    pub version: i64,
     /// Creation timestamp (UTC).
     pub created_at: DateTime<Utc>,
     /// Last-modification timestamp (UTC).
